- Fix OSR test failures (issue #1245). - Fix Mac and Windows compile errors.
git-svn-id: https://chromiumembedded.googlecode.com/svn/branches/1916@1715 5089003a-bbd8-11dd-ad1f-f1f9622dbc98

+++ tests/cefclient/cefclient_osr_dragdrop_win.cpp
@@ -252,7 +252,7 @@ bool DragDataToDataObject(CefRefPtr<CefDragData> drag_data,
     CefRefPtr<CefStreamWriter> writer =
         CefStreamWriter::CreateForHandler(handler.get());
     drag_data->GetFileContents(writer);
-    ASSERT(handler->GetDataSize() == bufferSize);
+    ASSERT(handler->GetDataSize() == static_cast<int64>(bufferSize));
     CefString fileName = drag_data->GetFileName();
     GetStorageForFileDescriptor(&stgmeds[curr_index], fileName.ToWString());
     fmtetc.cfFormat = file_desc_format;
